Project_Main/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==24.2.0
bidict==0.23.1
bleach==6.2.0
blinker==1.8.2
cachelib==0.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.3
dateparser==1.2.0
Deprecated==1.2.14
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
flask-marshmallow==1.2.1
Flask-Migrate==4.0.7
Flask-Session==0.8.0
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
frozenlist==1.4.1
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
msgspec==0.18.6
multidict==6.1.0
ordered-set==4.1.0
packaging==24.1
propcache==0.2.0
pycparser==2.22
pycryptodome==3.21.0
Pygments==2.18.0
PyJWT==2.9.0
python-binance==1.0.19
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
regex==2024.9.11
requests==2.32.3
rich==13.9.4
simple-websocket==1.1.0
six==1.16.0
SQLAlchemy==2.0.34
typing_extensions==4.12.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
webencodings==0.5.1
websockets==13.1
Werkzeug==3.0.4
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.2.1
yarl==1.15.4
zope.event==5.0
zope.interface==7.1.1
celery==5.4.0
redis==5.0.8
flask_celeryext
gunicorn